+++ src/view/event-add-button.js
@@ -2,7 +2,7 @@ import { createElement } from '../render.js';
 
 const createEventAddBtnTemplate = () => '<button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>';
 
-export default class EventAddBtn {
+export default class EventAddButton {
   getTemplate() {
     return createEventAddBtnTemplate();
   }
