+++ src/view/form-edit-point.js
@@ -1,52 +1,5 @@
 import { createElement } from '../render.js';
-
-const TRAVEL_TYPES = [
-  'Taxi',
-  'Bus',
-  'Train',
-  'Ship',
-  'Drive',
-  'Flight',
-  'Check-in',
-  'Sightseeing',
-  'Restaurant',
-];
-
-const TRAVEL_CITIES = [
-  'Amsterdam',
-  'Geneva',
-  'Chamonix',
-];
-
-const TRAVEL_OFFERS = [
-  {
-    key: 'luggage',
-    title: 'Add luggage',
-    price: '30',
-    checked: true,
-  },
-  {
-    key: 'comfort',
-    title: 'Switch to comfort class',
-    price: '100',
-    checked: true,
-  },
-  {
-    key: 'meal',
-    title: 'Add meal',
-    price: '15',
-  },
-  {
-    key: 'seats',
-    title: 'Choose seats',
-    price: '5',
-  },
-  {
-    key: 'train',
-    title: 'Travel by train',
-    price: '40',
-  }
-];
+import { TRAVEL_TYPES, TRAVEL_CITIES, TRAVEL_OFFERS } from '../constants.js';
 
 const createTravelTypeTemplate = (type) => `
   <div class="event__type-item">
