+++ src/view/trip-sort.js
@@ -1,36 +1,16 @@
 import { createElement } from '../render.js';
+import { TRAVEL_SORT_ITEMS } from '../constants.js';
+
+const createTripSortItemTemplate = (data) => `
+  <div class="trip-sort__item  trip-sort__item--${data.key}">
+    <input id="sort-${data.key}" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort"
+      value="sort-${data.key}" ${data.checked ? 'checked' : ''}  ${data.disabled ? 'disabled' : ''}>
+    <label class="trip-sort__btn" for="sort-${data.key}">${data.title}</label>
+  </div>`;
 
 const createTripSortTemplate = () => `
   <form class="trip-events__trip-sort  trip-sort" action="#" method="get">
-    <div class="trip-sort__item  trip-sort__item--day">
-      <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort"
-        value="sort-day" checked>
-      <label class="trip-sort__btn" for="sort-day">Day</label>
-    </div>
-
-    <div class="trip-sort__item  trip-sort__item--event">
-      <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort"
-        value="sort-event" disabled>
-      <label class="trip-sort__btn" for="sort-event">Event</label>
-    </div>
-
-    <div class="trip-sort__item  trip-sort__item--time">
-      <input id="sort-time" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort"
-        value="sort-time">
-      <label class="trip-sort__btn" for="sort-time">Time</label>
-    </div>
-
-    <div class="trip-sort__item  trip-sort__item--price">
-      <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort"
-        value="sort-price">
-      <label class="trip-sort__btn" for="sort-price">Price</label>
-    </div>
-
-    <div class="trip-sort__item  trip-sort__item--offer">
-      <input id="sort-offer" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort"
-        value="sort-offer" disabled>
-      <label class="trip-sort__btn" for="sort-offer">Offers</label>
-    </div>
+    ${TRAVEL_SORT_ITEMS.map((item) => createTripSortItemTemplate(item)).join('')}
   </form>`;
 
 export default class TripSort {
